pyproject.toml
[project]
name = "to-do-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=5.2.7",
    "djangorestframework>=3.16.1",
    "drf-yasg>=1.21.11",
    "gunicorn>=23.0.0",
    "python-decouple>=3.8",
    "whitenoise>=6.11.0",
]
